requirements.txt
fair-esm==2.0.0
einops==0.8.1
pdbparser==0.1.8
biopython==1.85
nglview==4.0
mmseqs==1.0.0

scikit-learn==1.7.2
scipy==1.16.2
pandas==1.5.3
numpy==1.26.4

torch==2.5.1
torchvision==0.20.1

easydict==1.13
matplotlib==3.9.2
Pillow==11.0.0
PyYAML==6.0.2
path==17.0.0
tqdm==4.67.1
pipreqs==0.2.3
